+++ Migrations/1.0/create_trucks_maintenance_table.php
@@ -1,5 +1,5 @@
 <?php
-namespace Modules\Brookr\Migrations\v1_0;
+namespace Modules\Brookr\Migrations;
 
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
@@ -9,19 +9,23 @@ class CreateTrucksMaintenanceTable extends Migration
 {
     public function up()
     {
-        Schema::create( 'brookr_trucks_maintenance', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string( 'name' );
-            $table->text( 'description' )->nullable();
-            $table->integer( 'truck_id' );
-            $table->text( 'document_url' );
-            $table->integer( 'user_id' );
-            $table->timestamps();
-        });
+        if ( ! Schema::hasTable( 'brookr_trucks_maintenance' ) ) {
+            Schema::create( 'brookr_trucks_maintenance', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->string( 'name' );
+                $table->text( 'description' )->nullable();
+                $table->integer( 'truck_id' );
+                $table->text( 'document_url' );
+                $table->integer( 'user_id' );
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
     {
-        Schema::drop( 'brookr_trucks_maintenance' );
+        if ( Schema::hasTable( 'brookr_trucks_maintenance' ) ) {
+            Schema::drop( 'brookr_trucks_maintenance' );
+        }
     }
 }
